invalid data error for input using custom hook
