+++ src/App.jsx
@@ -1,39 +1,19 @@
 import QRCode from "qrcode";
-import logo from "./assets/technigo-logo.svg";
-import { useState } from "react";
-export const App = () => {
-  // Destructure variables, properties and methods from the useQRCodeGenerator hook that you imported above here :)
-  // Define state variables for URL and QR code
-  const [url, setUrl] = useState("")
-  const [qr, setQr] = useState("")
+import { useQRCodeGenerator } from "./hooks/useQRCodeGenerator";
 
-  // Define a function to generate the QR code
-  const generateQRCode = () => {
-    QRCode.toDataURL(
-      url,
-      {
-        width: 100,
-        margin: 2,
-        color: {
-          dark: "#335383FF",
-          light: "#EEEEEEFF",
-        },
-      },
-      (err, url) => {
-        if (err) return console.error(err)
-        console.log(url)
-        setQr(url)
-      }
-    )
-  }
+export const App = () => {
+  // Initialize state variables and functions from useQRCodeGenerator
+  const {
+    url,
+    qr,
+    generateQRCode,
+    setUrl,
+  } = useQRCodeGenerator()
+
+
   
   return (
     <div className="">
-      {/* Render the title */}
-      {/* <img className="logo" src={logo} alt="" /> */}
-      {/* <h1>Technigo QR Code Generator</h1> */}
-      {/* <p>Start Here</p> */}
-      {/* <QrExample /> */}
       <h1>QR Generator</h1>
       <input 
         type="text"
