+++ src/hooks/useQRCodeGenerator copy.jsx
@@ -1,60 +1,24 @@
+// from technigo repo
 // Hook Explanation
 // This React component, specifically a custom hook named useQRCodeGenerator, is designed to facilitate the generation and downloading of QR codes. Initially, it utilizes the useState hook from React to manage three pieces of state: url (to store the input URL that will be converted into a QR code), qr (to store the generated QR code data URL), and showInput (a boolean to toggle the visibility of an input element in the UI). The hook exposes a method generateQRCode which utilizes the QRCode.toDataURL method to convert the provided URL into a QR code, applying specific styling options, and then updates the state with the generated QR code and hides the input. The downloadQRCode method allows users to download the generated QR code as a PNG file, prompting them to provide a filename and handling the download process via creating an anchor element in the DOM. Lastly, the repeatAction method resets the state to allow users to generate a new QR code. The hook returns an object containing the state variables and methods, enabling them to be utilized in the component where the hook is used.
 import QRCode from "qrcode";
-import { useState } from "react";
 
 // Define a custom hook named useQRCodeGenerator
 export const useQRCodeGenerator = () => {
   // Reactive State variable to store the input URL
   //   const ...
-  // Initialize url state
-  const [url, setUrl] = useState("")
 
   // Reactive State variable to store the generated QR code data URL
   //   const ...
-  // Initialize qr state
-  const [qr, setQr] = useState("")
 
   // Reactive State variable to toggle the visibility of the input element - boolean value
   //   const ...
-  // Initialize showInput
-  const [showInput, setShowInput] = useState(true)
 
   // Function to generate a QR code from the input URL
   const generateQRCode = () => {
-    // Your code for generating QR code goes here
-
-    // Update url state using setUrl
-    setUrl("new-url-value");
-
-    // Update qr state using setQr
-    setQr("new-qr-value");
-
-    // Update showInput state using setShowInput
-    setShowInput(false);
-
-
-
     // HINT 1: Utilize the qrccode library that converts a URL to a QR code data URL.
     // Use the Import of the qrcode and chain to the native method toDataUrl() much like the example provided and specify the data within the object that you will be passing such as the {url, {width, margin, color:{dark, light}}} which containes the information to generate the qr-code and url. Lastly, this native method toDataUrl() will contain a callback function  that will update the qr variable and will also update the variable toggling the visibility of the input element.
     QRCode.toDataURL(
-      url,
-      {
-        width: 100,
-        margin: 2,
-        color: {
-          dark: "#335383FF",
-          light: "#EEEEEEFF",
-        },
-      },
-      (err, generateQRCode) => {
-        if (err) {
-          console.error(err)
-        } else {
-          // update qr
-          setQr(generateQRCode)
-        }
-      },
       console.log("Delete This Line OR Comment Out")
       // HINT 2: Ensure to pass the necessary parameters to the QR code generation method, such as the URL to convert and any styling options.
       // ...
@@ -111,11 +75,5 @@ export const useQRCodeGenerator = () => {
   };
 
   // Return the state variables and functions to be used in the component
-  return {
-    url,
-    qr,
-    showInput,
-    generateQRCode,
-    setUrl,
-  };
+  return {};
 };
